requirements.txt
Pillow>=9.0.0
openai>=0.27.0
python-magic>=0.4.24
watchdog>=2.1.6
numpy>=1.21.0
pandas>=1.3.0